Move integration test to separate file.
